+++ tests/unit/controllers/new-platform-flexberry-services-lock-list-test.js
@@ -1,0 +1,8 @@
+import { moduleFor, test } from 'ember-qunit';
+
+moduleFor('controller:new-platform-flexberry-services-lock-list', 'Unit | Controller | new-platform-flexberry-services-lock-list');
+
+test('it exists', function(assert) {
+  let controller = this.subject();
+  assert.ok(controller);
+});